Eliminate duplicate code.
Eliminate duplicate code.
