+++ compiler/rustc/Cargo.toml
@@ -10,16 +10,7 @@ rustc_driver = { path = "../rustc_driver" }
 # crate is intended to be used by codegen backends, which may not be in-tree.
 rustc_codegen_ssa = { path = "../rustc_codegen_ssa" }
 
-[dependencies.tikv-jemalloc-sys]
-version = '0.4.0'
-optional = true
-features = ['unprefixed_malloc_on_supported_platforms']
-
-[dependencies.tikv-jemallocator]
-version = '0.4.0'
-optional = true
-
 [features]
-jemalloc = ['tikv-jemalloc-sys', 'tikv-jemallocator']
+jemalloc = ['rustc_driver/jemalloc']
 llvm = ['rustc_driver/llvm']
 max_level_info = ['rustc_driver/max_level_info']
