+++ src/tools/tidy/src/deps.rs
@@ -230,6 +230,9 @@ const PERMITTED_DEPENDENCIES: &[&str] = &[
     "winapi-x86_64-pc-windows-gnu",
     // this is a false-positive: it's only used by rustfmt, but because it's enabled through a feature, tidy thinks it's used by rustc as well.
     "yansi-term",
+    "tikv-jemalloc-sys",
+    "tikv-jemallocator",
+    "fs_extra",
 ];
 
 const PERMITTED_CRANELIFT_DEPENDENCIES: &[&str] = &[
